pollpps: drive the serialpps backend instead of a private poll loop
pollpps hand-rolled a 10 kHz ticker poll and its own edge detection,
a parallel implementation of what time/internal/serialpps now does
better. Move the tool's logic to a new time/app/pollppscmd package,
inside the time tree so it can import serialpps, leaving cmd/pollpps
a thin main. The monitoring mode now runs serialpps.Poll, echoing its
debug log to stderr, so the tool exercises the daemon's actual edge
detection path; the ioctl timing mode moves into serialpps as
TimeStateReads, next to the polling loop whose pacing floor it
measures, where it uses the same clock readings as the loop.

The package is provisional scaffolding: equivalent functionality is
expected to become a satpulsetool subcommand once its shape is
decided.
